Fix some errors, move TerminalCore to common/
